test: Extract config tests and drop unsupported syntax
- Removes the `Envs(test, uat)` parenthesis syntax as it is unnecessary.
- Extracts `TestCustomEnvConfig` and `TestFindConfig` into `config_test.go` to cleanly separate testing responsibilities.
