+++ tag_test.go
@@ -449,60 +449,3 @@ func TestTag_Clone(t *testing.T) {
 		}
 	})
 }
-
-func TestCustomEnvConfig(t *testing.T) {
-	// Override configuration
-	parseTagReLock.Lock()
-	originalEnvs := ConfiguredEnvs
-	originalMap := ConfiguredEnvsMap
-	parseTagReLock.Unlock()
-
-	err := LoadConfig("testdata/test_config.conf")
-	if err != nil {
-		t.Fatalf("failed to load config: %v", err)
-	}
-
-	defer func() {
-		parseTagReLock.Lock()
-		ConfiguredEnvs = originalEnvs
-		ConfiguredEnvsMap = originalMap
-		parseTagRe = nil
-		parseTagReLock.Unlock()
-	}()
-
-	// Parse test
-	tag := ParseTag("v1.0.0-staging1")
-	if tag == nil || tag.EnvName != "staging" || *tag.Env != 1 {
-		t.Fatalf("failed to parse custom env: %+v", tag)
-	}
-
-	// Format test
-	if got := tag.String(); got != "v1.0.0-staging1" {
-		t.Fatalf("expected v1.0.0-staging1, got %s", got)
-	}
-
-	// CommandsToFlags test
-	flags := CommandsToFlags([]string{"prod2"}, "default")
-	if flags.Env != "prod" || *flags.EnvValue != 2 {
-		t.Fatalf("failed to parse custom env flag: %+v", flags)
-	}
-
-	// Increment test
-	err = tag.Increment(flags, false, false)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if got := tag.String(); got != "v1.0.0-prod02" {
-		t.Fatalf("expected v1.0.0-prod02, got %s", got)
-	}
-}
-
-func TestFindConfig(t *testing.T) {
-	path, err := FindConfig("testdata/test_config.conf")
-	if err != nil {
-		t.Fatalf("expected to find config, got %v", err)
-	}
-	if path == "" {
-		t.Fatalf("expected path to not be empty")
-	}
-}
