--- testdata/test_config_paren.conf
@@ -1,2 +1,0 @@
-# This is a test config file
-Envs(staging, prod)
